refactor: moved custom Link to a new shared component for reusability
