+++ pages/index.tsx
@@ -1,9 +1,8 @@
 import Head from 'next/head';
 import Image from 'next/image';
 import styles from '../styles/Home.module.css';
-import Link from 'next/link';
 import { css } from '@emotion/react';
-import { ReactNode } from 'react';
+import Link from '../src/components/Link';
 
 const Home = () => {
   return (
@@ -18,14 +17,14 @@ const Home = () => {
         <h1 className={styles.title}>Welcome to .SHIFT habit tracker!</h1>
 
         <p>
-          <StyledLink href={`/about`}>About the app</StyledLink>
+          <Link href={`/about`}>About the app</Link>
         </p>
 
         <p className={styles.links}>
-          <StyledLink href={`/profile/Manpreet`}>Manpreet Singh</StyledLink>
-          <StyledLink href={`profile/Harjot`}>Harjot Singh</StyledLink>
-          <StyledLink href={`profile/Gurpreet`}>Gurpreet Kaur</StyledLink>
-          <StyledLink href={`/profile/Hardeep`}>Hardeep Kaur</StyledLink>
+          <Link href={`/profile/Manpreet`}>Manpreet Singh</Link>
+          <Link href={`profile/Harjot`}>Harjot Singh</Link>
+          <Link href={`profile/Gurpreet`}>Gurpreet Kaur</Link>
+          <Link href={`/profile/Hardeep`}>Hardeep Kaur</Link>
           <br />
         </p>
       </main>
@@ -54,23 +53,3 @@ const containerStyle = css({
   color: '#fff',
   textAlign: 'center'
 });
-
-type StyledLinkProps = {
-  href: string;
-  children: ReactNode;
-};
-
-const StyledLink = ({ href, children }: StyledLinkProps) => {
-  const linkStyle = css({
-    fontSize: '1.2em',
-    '&:hover': {
-      color: 'orange'
-    }
-  });
-
-  return (
-    <Link css={linkStyle} href={href}>
-      <>{children}</>
-    </Link>
-  );
-};
