+++ src/components/Link.tsx
@@ -1,0 +1,25 @@
+import NextLink from 'next/link';
+import { ReactNode } from 'react';
+import { css } from '@emotion/react';
+
+type LinkProps = {
+  href: string;
+  children: ReactNode;
+};
+
+const Link = ({ href, children }: LinkProps) => {
+  const linkStyle = css({
+    fontSize: '1.2em',
+    '&:hover': {
+      color: 'orange'
+    }
+  });
+
+  return (
+    <NextLink css={linkStyle} href={href}>
+      <>{children}</>
+    </NextLink>
+  );
+};
+
+export default Link;
